# db: apply trySeekUsingNext optimization to file metadata iterator